+++ game/missiongenerator/aircraft/waypoints/landingpoint.py
@@ -1,5 +1,6 @@
 from dcs.point import MovingPoint, PointAction
 
+from game.theater import NavalControlPoint
 from .pydcswaypointbuilder import PydcsWaypointBuilder
 
 
@@ -9,5 +10,9 @@ class LandingPointBuilder(PydcsWaypointBuilder):
         waypoint.type = "Land"
         waypoint.action = PointAction.Landing
         if (control_point := self.waypoint.control_point) is not None:
-            waypoint.airdrome_id = control_point.airdrome_id_for_landing
+            if isinstance(control_point, NavalControlPoint):
+                waypoint.helipad_id = control_point.airdrome_id_for_landing
+                waypoint.link_unit = control_point.airdrome_id_for_landing
+            else:
+                waypoint.airdrome_id = control_point.airdrome_id_for_landing
         return waypoint